transaction/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==1.0.0
Django==4.0.3
django-bootstrap3==21.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-tables2==2.4.1
-e git+https://github.com/sghosh28/django-tenant-schemas.git@ef4964bb6b3457e309190b9367c9dbed5b3b8548#egg=django_tenant_schemas
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
ordered-set==4.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
toml==0.10.2
tzdata==2021.5
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
